102 fibonacci numbers

+++ functions_nested_loops/102-fibonacci.c
@@ -6,8 +6,8 @@
 int main(void)
 {
 int i = 0;
-int j = 1;
-int k = 2;
+long j = 1;
+long k = 2;
 while (i < 50)
 {
 if (i == 0)
